Function to download financial data

This function downloads financial data from Yahoo Finance using getSymbols. Based on a set of tickers and a time period, the function will download the data for each ticker and return a report of the process, along with the actual data in the long dataframe format. The main advantage of the function is that it automatically recognizes the source of the dataset from the ticker and structures the resulting data from different sources in the long format. A caching system is also available, making it very fast.

BatchGetSymbols(
  tickers,
  first.date = Sys.Date() - 30,
  last.date = Sys.Date(),
  thresh.bad.data = 0.75,
  bench.ticker = "^GSPC",
  type.return = "arit",
  freq.data = "daily",
  how.to.aggregate = "last",
  do.complete.data = FALSE,
  do.fill.missing.prices = TRUE,
  do.cache = TRUE,
  cache.folder = file.path(tempdir(), "BGS_Cache"),
  do.parallel = FALSE,
  be.quiet = FALSE
)

Arguments

tickers

A vector of tickers. If not sure whether the ticker is available, check the websites of google and yahoo finance. The source for downloading the data can either be Google or Yahoo. The function automatically selects the source webpage based on the input ticker.

first.date

The first date to download data (date or char as YYYY-MM-DD)

last.date

The last date to download data (date or char as YYYY-MM-DD)

thresh.bad.data

A percentage threshold for defining bad data. The dates of the benchmark ticker are compared to each asset. If the percentage of non-missing dates with respect to the benchmark ticker is lower than thresh.bad.data, the function will ignore the asset (default = 0.75)

bench.ticker

The ticker of the benchmark asset used to compare dates. My suggestion is to use the main stock index of the market from where the data is coming from (default = ^GSPC (SP500, US market))

type.return

Type of price return to calculate: 'arit' (default) - aritmetic, 'log' - log returns.

freq.data

Frequency of financial data ('daily', 'weekly', 'monthly', 'yearly')

how.to.aggregate

Defines whether to aggregate the data using the first observations of the aggregating period or last ('first', 'last'). For example, if freq.data = 'yearly' and how.to.aggregate = 'last', the last available day of the year will be used for all aggregated values such as price.adjusted.

do.complete.data

Return a complete/balanced dataset? If TRUE, all missing pairs of ticker-date will be replaced by NA or closest price (see input do.fill.missing.prices). Default = FALSE.

do.fill.missing.prices

Finds all missing prices and replaces them by their closest price with preference for the previous price. This ensures a balanced dataset for all assets, without any NA. Default = TRUE.

do.cache

Use cache system? (default = TRUE)

cache.folder

Where to save cache files? (default = file.path(tempdir(), 'BGS_Cache') )

do.parallel

Flag for using parallel or not (default = FALSE). Before using parallel, make sure you call function future::plan() first.

be.quiet

Logical for printing statements (default = FALSE)

Value

A list with the following items:

df.control

A dataframe containing the results of the download process for each asset

df.tickers

A dataframe with the financial data for all valid tickers

Warning

Do notice that since 2019, adjusted prices are no longer available from google finance. When using this source, the function will output NA values for this column.

Also, be aware that when using cache system in a local folder (and not the default tempdir()), the aggregate prices series might not match if a split or dividends event happens in between cache files.

See also

Examples

tickers <- c('FB','MMM')

first.date <- Sys.Date()-30
last.date <- Sys.Date()

l.out <- BatchGetSymbols(tickers = tickers,
                         first.date = first.date,
                        last.date = last.date, do.cache=FALSE)
#> 
#> Running BatchGetSymbols for:
#> 
#>    tickers =FB, MMM
#> 
#>    Downloading data for benchmark ticker
#> 
#> ^GSPC | yahoo (1|1)
#> 
#> FB | yahoo (1|2)
#>  - Got 100% of valid prices | Good job!
#> 
#> MMM | yahoo (2|2)
#>  - Got 100% of valid prices | Mais faceiro que guri de bombacha nova!

print(l.out$df.control)
#> # A tibble: 2 × 6
#>   ticker src   download.status total.obs perc.benchmark.dates threshold.decision
#>   <chr>  <chr> <chr>               <int>                <dbl> <chr>             
#> 1 FB     yahoo OK                     22                    1 KEEP              
#> 2 MMM    yahoo OK                     22                    1 KEEP              
print(l.out$df.tickers)
#>    price.open price.high price.low price.close   volume price.adjusted
#> 1      207.68     213.15    206.83      211.03 34239800         211.03
#> 2      209.87     211.75    202.22      203.49 27094900         203.49
#> 3      204.78     209.15    202.02      208.11 29452100         208.11
#> 4      209.02     209.08    201.42      202.97 27263500         202.97
#> 5      202.37     205.89    198.79      200.06 32130900         200.06
#> 6      200.79     200.89    187.28      187.47 38560600         187.47
#> 7      187.85     196.70    186.11      190.29 37508100         190.29
#> 8      195.74     199.10    193.92      198.50 31894700         198.50
#> 9      194.84     196.28    191.07      195.21 24853000         195.21
#> 10     192.64     193.56    186.67      187.61 34656900         187.61
#> 11     187.03     192.29    185.82      186.63 31010500         186.63
#> 12     190.80     192.28    185.91      192.03 31721700         192.03
#> 13     194.70     203.73    194.52      203.63 40640300         203.63
#> 14     201.70     208.04    201.11      207.84 29499700         207.84
#> 15     206.70     216.80    206.00      216.49 52055000         216.49
#> 16     214.50     214.71    207.63      211.49 30142300         211.49
#> 17     211.37     219.46    210.18      216.65 31998800         216.65
#> 18     213.33     216.80    212.16      213.46 23717300         213.46
#> 19     215.00     220.67    214.79      219.57 31502300         219.57
#> 20     220.51     225.50    218.89      221.82 40010500         221.82
#> 21     222.13     224.04    219.54      223.59 26224100         223.59
#> 22     226.07     230.89    225.29      229.86 31391200         229.86
#> 23     149.29     149.63    146.75      148.65  4214800         148.65
#> 24     148.58     149.02    144.15      144.75  3313900         144.75
#> 25     145.61     148.37    144.92      147.22  3734600         147.22
#> 26     148.52     149.99    147.52      148.30  3804200         148.30
#> 27     146.16     147.69    145.74      146.73  3235200         146.73
#> 28     145.74     145.74    142.57      143.28  4554000         143.28
#> 29     145.24     149.47    143.26      145.40  4605300         145.40
#> 30     147.86     149.26    146.49      146.60  3337300         146.60
#> 31     145.34     145.96    143.33      143.93  3794000         143.93
#> 32     143.47     144.48    140.83      140.96  4148700         140.96
#> 33     141.45     145.29    141.45      143.48  3910100         143.48
#> 34     144.87     145.48    142.59      144.31  3337300         144.31
#> 35     144.76     146.27    142.84      145.51  3813800         145.51
#> 36     145.31     147.71    144.15      147.69  2786700         147.69
#> 37     147.07     148.45    145.22      148.32  7881800         148.32
#> 38     148.69     149.74    147.40      148.58  4514900         148.58
#> 39     149.45     150.84    148.50      149.94  4252000         149.94
#> 40     149.69     149.78    147.56      147.68  2637200         147.68
#> 41     148.12     149.23    147.40      148.98  2028100         148.98
#> 42     149.27     151.18    148.91      150.46  2247800         150.46
#> 43     149.93     150.13    147.19      149.47  2679900         149.47
#> 44     151.00     152.92    150.04      152.08  2692300         152.08
#>      ref.date ticker ret.adjusted.prices ret.closing.prices
#> 1  2022-02-28     FB                  NA                 NA
#> 2  2022-03-01     FB        -0.035729489       -0.035729489
#> 3  2022-03-02     FB         0.022703798        0.022703798
#> 4  2022-03-03     FB        -0.024698477       -0.024698477
#> 5  2022-03-04     FB        -0.014337109       -0.014337109
#> 6  2022-03-07     FB        -0.062931106       -0.062931106
#> 7  2022-03-08     FB         0.015042364        0.015042364
#> 8  2022-03-09     FB         0.043144712        0.043144712
#> 9  2022-03-10     FB        -0.016574272       -0.016574272
#> 10 2022-03-11     FB        -0.038932461       -0.038932461
#> 11 2022-03-14     FB        -0.005223581       -0.005223581
#> 12 2022-03-15     FB         0.028934222        0.028934222
#> 13 2022-03-16     FB         0.060407260        0.060407260
#> 14 2022-03-17     FB         0.020674709        0.020674709
#> 15 2022-03-18     FB         0.041618597        0.041618597
#> 16 2022-03-21     FB        -0.023095754       -0.023095754
#> 17 2022-03-22     FB         0.024398264        0.024398264
#> 18 2022-03-23     FB        -0.014724150       -0.014724150
#> 19 2022-03-24     FB         0.028623629        0.028623629
#> 20 2022-03-25     FB         0.010247301        0.010247301
#> 21 2022-03-28     FB         0.007979393        0.007979393
#> 22 2022-03-29     FB         0.028042422        0.028042422
#> 23 2022-02-28    MMM                  NA                 NA
#> 24 2022-03-01    MMM        -0.026236086       -0.026236086
#> 25 2022-03-02    MMM         0.017063910        0.017063910
#> 26 2022-03-03    MMM         0.007335973        0.007335973
#> 27 2022-03-04    MMM        -0.010586696       -0.010586696
#> 28 2022-03-07    MMM        -0.023512554       -0.023512554
#> 29 2022-03-08    MMM         0.014796168        0.014796168
#> 30 2022-03-09    MMM         0.008253178        0.008253178
#> 31 2022-03-10    MMM        -0.018212912       -0.018212912
#> 32 2022-03-11    MMM        -0.020634935       -0.020634935
#> 33 2022-03-14    MMM         0.017877333        0.017877333
#> 34 2022-03-15    MMM         0.005784792        0.005784792
#> 35 2022-03-16    MMM         0.008315411        0.008315411
#> 36 2022-03-17    MMM         0.014981837        0.014981837
#> 37 2022-03-18    MMM         0.004265725        0.004265725
#> 38 2022-03-21    MMM         0.001752933        0.001752933
#> 39 2022-03-22    MMM         0.009153318        0.009153318
#> 40 2022-03-23    MMM        -0.015072756       -0.015072756
#> 41 2022-03-24    MMM         0.008802838        0.008802838
#> 42 2022-03-25    MMM         0.009934293        0.009934293
#> 43 2022-03-28    MMM        -0.006579861       -0.006579861
#> 44 2022-03-29    MMM         0.017461705        0.017461705