Function to download financial data

This function is designed to make batch downloads of financial data using getSymbols. Based on a set of tickers and a time period, the function will download the data for each ticker and return a report of the process, along with the actual data in the long dataframe format. The main advantage of the function is that it automatically recognizes the source of the dataset from the ticker and structures the resulting data from different sources in the long format. A caching system is also presente, making it very fast.

BatchGetSymbols(tickers, first.date = Sys.Date() - 30,
  last.date = Sys.Date(), thresh.bad.data = 0.75,
  bench.ticker = "^GSPC", type.return = "arit", freq.data = "daily",
  do.complete.data = FALSE, do.cache = TRUE,
  cache.folder = "BGS_Cache")

Arguments

tickers
A vector of tickers. If not sure whether the ticker is available, check the websites of google and yahoo finance. The source for downloading the data can either be Google or Yahoo. The function automatically selects the source webpage based on the input ticker.

first.date
The first date to download data (date or char as YYYY-MM-DD)

last.date
The last date to download data (date or char as YYYY-MM-DD)

thresh.bad.data
A percentage threshold for defining bad data. The dates of the benchmark ticker are compared to each asset. If the percentage of non-missing dates with respect to the benchmark ticker is lower than thresh.bad.data, the function will ignore the asset (default = 0.75)

bench.ticker
The ticker of the benchmark asset used to compare dates. My suggestion is to use the main stock index of the market from where the data is coming from (default = ^GSPC (SP500, US market))

type.return
Type of price return to calculate: 'arit' (default) - aritmetic, 'log' - log returns.

freq.data
Frequency of financial data ('daily', 'weekly', 'monthly', 'yearly')

do.complete.data
Return a complete/balanced dataset? If TRUE, all missing pairs of ticker-date will be replaced by NA (default = FALSE)

do.cache
Use caching system? (default = TRUE)

cache.folder
Where to save cache files? (default = 'BGS_Cache')

Value

A list with the following items:

df.control

A dataframe containing the results of the download process for each asset

df.tickers

A dataframe with the financial data for all valid tickers

Warning

Do notice that adjusted prices are not available from google finance. When using this source, the function will output NA values for this column.

See also

Examples

tickers <- c('FB','NYSE:MMM')

first.date <- Sys.Date()-30
last.date <- Sys.Date()

l.out <- BatchGetSymbols(tickers = tickers,
                         first.date = first.date,
                        last.date = last.date, do.cache=FALSE)

#> 
#> Running BatchGetSymbols for:
#>    tickers = FB, NYSE:MMM
#>    Downloading data for benchmark ticker
#> FB | yahoo (1|2) - OK!
#> NYSE:MMM | google (2|2) - Error in download..

print(l.out$df.control)

#>     ticker    src download.status total.obs perc.benchmark.dates
#> 1       FB  yahoo              OK        22                    1
#> 2 NYSE:MMM google          NOT OK         0                    0
#>   threshold.decision
#> 1               KEEP
#> 2                OUT

print(l.out$df.tickers)

#>    price.open price.high price.low price.close   volume price.adjusted
#> 1      161.58     162.28    158.67      159.33 36031000         159.33
#> 2      160.00     163.66    159.53      162.43 23109500         162.43
#> 3      161.46     161.46    157.35      158.85 25739600         158.85
#> 4      159.21     160.90    156.20      157.33 25744000         157.33
#> 5      155.54     158.34    154.39      157.25 24046000         157.25
#> 6      157.69     160.59    157.42      157.90 18844400         157.90
#> 7      156.82     157.69    151.31      151.38 30610000         151.38
#> 8      150.13     154.81    149.16      153.35 35338900         153.35
#> 9      156.73     156.89    151.30      153.74 25293500         153.74
#> 10     153.32     155.57    152.55      153.52 15433500         153.52
#> 11     155.40     159.46    155.01      158.78 19180100         158.78
#> 12     159.56     160.49    157.95      159.42 17592000         159.42
#> 13     158.51     158.66    153.28      154.92 21675100         154.92
#> 14     155.86     157.35    153.55      154.05 19761300         154.05
#> 15     154.76     157.34    154.46      154.78 15424700         154.78
#> 16     151.22     154.77    150.85      154.39 19095000         154.39
#> 17     154.28     154.65    145.60      146.04 27744600         146.04
#> 18     147.73     152.21    147.00      150.95 22105700         150.95
#> 19     145.82     149.00    143.80      145.37 31303300         145.37
#> 20     148.50     148.83    139.03      142.09 31336800         142.09
#> 21     139.94     146.64    139.74      146.22 50528300         146.22
#> 22     155.00     156.40    148.96      151.79 59993400         151.79
#>      ref.date ticker ret.adjusted.prices ret.closing.prices
#> 1  2018-10-02     FB                  NA                 NA
#> 2  2018-10-03     FB        0.0194564173       0.0194564173
#> 3  2018-10-04     FB       -0.0220401844      -0.0220401844
#> 4  2018-10-05     FB       -0.0095688004      -0.0095688004
#> 5  2018-10-08     FB       -0.0005084981      -0.0005084981
#> 6  2018-10-09     FB        0.0041335072       0.0041335072
#> 7  2018-10-10     FB       -0.0412918888      -0.0412918888
#> 8  2018-10-11     FB        0.0130136143       0.0130136143
#> 9  2018-10-12     FB        0.0025431952       0.0025431952
#> 10 2018-10-15     FB       -0.0014309938      -0.0014309938
#> 11 2018-10-16     FB        0.0342626033       0.0342626033
#> 12 2018-10-17     FB        0.0040307281       0.0040307281
#> 13 2018-10-18     FB       -0.0282273244      -0.0282273244
#> 14 2018-10-19     FB       -0.0056157695      -0.0056157695
#> 15 2018-10-22     FB        0.0047386951       0.0047386951
#> 16 2018-10-23     FB       -0.0025197054      -0.0025197054
#> 17 2018-10-24     FB       -0.0540838529      -0.0540838529
#> 18 2018-10-25     FB        0.0336209548       0.0336209548
#> 19 2018-10-26     FB       -0.0369658967      -0.0369658967
#> 20 2018-10-29     FB       -0.0225631087      -0.0225631087
#> 21 2018-10-30     FB        0.0290661209       0.0290661209
#> 22 2018-10-31     FB        0.0380932291       0.0380932291