Function to download financial data

This function is designed to make batch downloads of financial data using getSymbols. Based on a set of tickers and a time period, the function will download the data for each ticker and return a report of the process, along with the actual data in the long dataframe format. The main advantage of the function is that it automatically recognizes the source of the dataset from the ticker and structures the resulting data from different sources in the long format. A caching system is also presente, making it very fast.

BatchGetSymbols(tickers, first.date = Sys.Date() - 30,
  last.date = Sys.Date(), thresh.bad.data = 0.75,
  bench.ticker = "^GSPC", type.return = "arit", freq.data = "daily",
  do.complete.data = FALSE, do.fill.missing.prices = TRUE,
  do.cache = TRUE, cache.folder = "BGS_Cache")

Arguments

tickers
A vector of tickers. If not sure whether the ticker is available, check the websites of google and yahoo finance. The source for downloading the data can either be Google or Yahoo. The function automatically selects the source webpage based on the input ticker.

first.date
The first date to download data (date or char as YYYY-MM-DD)

last.date
The last date to download data (date or char as YYYY-MM-DD)

thresh.bad.data
A percentage threshold for defining bad data. The dates of the benchmark ticker are compared to each asset. If the percentage of non-missing dates with respect to the benchmark ticker is lower than thresh.bad.data, the function will ignore the asset (default = 0.75)

bench.ticker
The ticker of the benchmark asset used to compare dates. My suggestion is to use the main stock index of the market from where the data is coming from (default = ^GSPC (SP500, US market))

type.return
Type of price return to calculate: 'arit' (default) - aritmetic, 'log' - log returns.

freq.data
Frequency of financial data ('daily', 'weekly', 'monthly', 'yearly')

do.complete.data
Return a complete/balanced dataset? If TRUE, all missing pairs of ticker-date will be replaced by NA or closest price (see input do.fill.missing.prices). Default = FALSE.

do.fill.missing.prices
Finds all missing prices and replaces them by their closest price with preference for the previous price. This ensures a balanced dataset for all assets, without any NA. Default = TRUE.

do.cache
Use caching system? (default = TRUE)

cache.folder
Where to save cache files? (default = 'BGS_Cache')

Value

A list with the following items:

df.control

A dataframe containing the results of the download process for each asset

df.tickers

A dataframe with the financial data for all valid tickers

Warning

Do notice that adjusted prices are not available from google finance. When using this source, the function will output NA values for this column.

See also

Examples

tickers <- c('FB','MMM')

first.date <- Sys.Date()-30
last.date <- Sys.Date()

l.out <- BatchGetSymbols(tickers = tickers,
                         first.date = first.date,
                        last.date = last.date, do.cache=FALSE)

#> 
#> Running BatchGetSymbols for:
#>    tickers = FB, MMM
#>    Downloading data for benchmark ticker
#> FB | yahoo (1|2) - Got 100% of valid prices | OK!
#> MMM | yahoo (2|2) - Got 100% of valid prices | You got it!

print(l.out$df.control)

#>   ticker   src download.status total.obs perc.benchmark.dates
#> 1     FB yahoo              OK        20                    1
#> 2    MMM yahoo              OK        20                    1
#>   threshold.decision
#> 1               KEEP
#> 2               KEEP

print(l.out$df.tickers)

#>    price.open price.high price.low price.close   volume price.adjusted
#> 1      145.82     149.00    143.80      145.37 31303300       145.3700
#> 2      148.50     148.83    139.03      142.09 31336800       142.0900
#> 3      139.94     146.64    139.74      146.22 50528300       146.2200
#> 4      155.00     156.40    148.96      151.79 60101300       151.7900
#> 5      151.52     152.75    149.35      151.75 25640800       151.7500
#> 6      151.80     154.13    148.96      150.35 24708700       150.3500
#> 7      150.10     150.19    147.44      148.68 15971200       148.6800
#> 8      149.31     150.97    148.00      149.94 16667100       149.9400
#> 9      151.57     153.01    149.83      151.53 21877400       151.5300
#> 10     150.49     150.94    146.74      147.87 24145800       147.8700
#> 11     146.75     147.76    144.07      144.96 17326900       144.9600
#> 12     144.48     145.04    140.49      141.55 18542100       141.5500
#> 13     142.00     144.88    141.62      142.16 15141700       142.1600
#> 14     143.70     145.58    141.55      144.22 22068400       144.2200
#> 15     142.33     144.84    140.83      143.85 30320300       143.8500
#> 16     141.07     141.77    137.77      139.53 37250600       139.5300
#> 17     137.61     137.75    131.21      131.55 44362700       131.5500
#> 18     127.03     134.16    126.85      132.43 41939500       132.4300
#> 19     134.40     137.19    134.13      134.82 25469700       134.8200
#> 20     133.65     134.50    131.26      131.73 11886100       131.7300
#> 21     183.31     186.66    181.98      184.95  3546200       183.6955
#> 22     187.85     188.88    183.16      185.79  2867600       184.5298
#> 23     186.83     190.08    185.00      189.81  2919600       188.5226
#> 24     190.70     192.12    189.67      190.26  2845100       188.9695
#> 25     190.61     193.21    190.09      192.22  2242800       190.9162
#> 26     193.66     194.47    191.60      192.81  3069600       191.5022
#> 27     193.50     194.56    192.21      193.25  2886500       191.9392
#> 28     193.27     195.50    193.00      195.30  1900200       193.9753
#> 29     197.19     201.40    195.41      201.29  2444400       199.9247
#> 30     201.00     204.27    200.73      204.07  2534000       202.6859
#> 31     204.00     204.00    199.70      201.10  2429300       199.7360
#> 32     200.57     200.77    196.50      197.04  2770000       195.7035
#> 33     197.52     200.88    197.16      198.37  2168800       197.0245
#> 34     200.80     201.04    196.29      198.06  2797700       196.7166
#> 35     197.73     205.13    196.24      204.91  3157500       203.5202
#> 36     203.24     209.77    202.60      209.00  2620400       207.5824
#> 37     207.83     208.05    204.37      205.58  3220400       204.1856
#> 38     203.58     204.14    200.19      200.51  2677200       199.1500
#> 39     199.68     201.18    198.66      198.69  1832700       198.6900
#> 40     197.76     198.68    196.79      196.80   999900       196.8000
#>      ref.date ticker ret.adjusted.prices ret.closing.prices
#> 1  2018-10-26     FB                  NA                 NA
#> 2  2018-10-29     FB       -0.0225631087      -0.0225631087
#> 3  2018-10-30     FB        0.0290661209       0.0290661209
#> 4  2018-10-31     FB        0.0380932291       0.0380932291
#> 5  2018-11-01     FB       -0.0002634759      -0.0002634759
#> 6  2018-11-02     FB       -0.0092256606      -0.0092256606
#> 7  2018-11-05     FB       -0.0111075021      -0.0111075021
#> 8  2018-11-06     FB        0.0084746372       0.0084746372
#> 9  2018-11-07     FB        0.0106042215       0.0106042215
#> 10 2018-11-08     FB       -0.0241536595      -0.0241536595
#> 11 2018-11-09     FB       -0.0196793677      -0.0196793677
#> 12 2018-11-12     FB       -0.0235237571      -0.0235237571
#> 13 2018-11-13     FB        0.0043094383       0.0043094383
#> 14 2018-11-14     FB        0.0144906932       0.0144906932
#> 15 2018-11-15     FB       -0.0025654902      -0.0025654902
#> 16 2018-11-16     FB       -0.0300313300      -0.0300313300
#> 17 2018-11-19     FB       -0.0571919735      -0.0571919735
#> 18 2018-11-20     FB        0.0066893955       0.0066893955
#> 19 2018-11-21     FB        0.0180473769       0.0180473769
#> 20 2018-11-23     FB       -0.0229195286      -0.0229195286
#> 21 2018-10-26    MMM                  NA                 NA
#> 22 2018-10-29    MMM        0.0045416998       0.0045417465
#> 23 2018-10-30    MMM        0.0216373525       0.0216373602
#> 24 2018-10-31    MMM        0.0023707772       0.0023707761
#> 25 2018-11-01    MMM        0.0103017464       0.0103017242
#> 26 2018-11-02    MMM        0.0030694038       0.0030693840
#> 27 2018-11-05    MMM        0.0022820204       0.0022820497
#> 28 2018-11-06    MMM        0.0106080549       0.0106080362
#> 29 2018-11-07    MMM        0.0306706818       0.0306707113
#> 30 2018-11-08    MMM        0.0138109700       0.0138109896
#> 31 2018-11-09    MMM       -0.0145537736      -0.0145538340
#> 32 2018-11-12    MMM       -0.0201890788      -0.0201890248
#> 33 2018-11-13    MMM        0.0067499245       0.0067499089
#> 34 2018-11-14    MMM       -0.0015627092      -0.0015627212
#> 35 2018-11-15    MMM        0.0345855028       0.0345855098
#> 36 2018-11-16    MMM        0.0199599689       0.0199599625
#> 37 2018-11-19    MMM       -0.0163636454      -0.0163636268
#> 38 2018-11-20    MMM       -0.0246619439      -0.0246619659
#> 39 2018-11-21    MMM       -0.0023097766      -0.0090768193
#> 40 2018-11-23    MMM       -0.0095123005      -0.0095123005