An improved version of function getSymbols from quantmod

This is a helper function to BatchGetSymbols and it should normally not be called directly. The purpose of this function is to download financial data based on a ticker and a time period. The main difference from getSymbols is that it imports the data as a dataframe with proper named columns and saves data locally with the caching system.

myGetSymbols(ticker, src = "yahoo", first.date, last.date,
  do.cache = TRUE, cache.folder = file.path(tempdir(), "BGS_Cache"))

Arguments

ticker
A single ticker to download data

src
The source of the data ('google' or'yahoo')

first.date
The first date to download data (date or char as YYYY-MM-DD)

last.date
The last date to download data (date or char as YYYY-MM-DD)

do.cache
Use caching system? (default = TRUE)

cache.folder
Where to save cache files? (default = 'BGS_Cache')

Value

A dataframe with the financial data

See also

Examples

ticker <- 'FB'

first.date <- Sys.Date()-30
last.date <- Sys.Date()
# NOT RUN {
df.ticker <- myGetSymbols(ticker,
                          first.date = first.date,
                          last.date = last.date)
# }