Transforms a dataframe in the long format to a list of dataframes in the wide format

Transforms a dataframe in the long format to a list of dataframes in the wide format

reshape.wide(df.tickers)

Arguments

df.tickers
Dataframe in the long format

Value

A list with dataframes in the wide format

Examples

my.f <- system.file( 'extdata/ExampleData.rds', package = 'BatchGetSymbols' )
df.tickers <- readRDS(my.f)
l.wide <- reshape.wide(df.tickers)
l.wide

#> $price.open
#> # A tibble: 248 x 16
#>    ref.date   ABEV3.SA BBAS3.SA BBDC3.SA BBDC4.SA BBSE3.SA BRAP4.SA BRFS3.SA
#>    <date>        <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#>  1 2014-01-02     17.3     24.4     21.9     20       24.5     25.0     48.8
#>  2 2014-01-03     17.4     24.0     21.6     19.9     24       24.2     48.5
#>  3 2014-01-06     16.9     24.0     21.1     19.6     23.5     23.8     47.2
#>  4 2014-01-07     17.0     23.9     21.2     19.8     23.1     23.9     47.1
#>  5 2014-01-08     17.3     23.5     21.0     19.6     23       23.4     47.2
#>  6 2014-01-09     17.5     23.0     20.8     19.3     23.1     23.1     45.7
#>  7 2014-01-10     17.1     22.3     20.2     18.9     22.6     22.2     44.8
#>  8 2014-01-13     17.2     23.0     20.6     19.1     22.9     22.3     44.3
#>  9 2014-01-14     17.0     22.7     20.3     18.8     23       22.1     44.5
#> 10 2014-01-15     17.1     23.4     20.6     19.1     23.5     22.4     44.8
#> # ... with 238 more rows, and 8 more variables: BRKM5.SA <dbl>, BRML3.SA <dbl>,
#> #   BVMF3.SA <dbl>, CCRO3.SA <dbl>, CIEL3.SA <dbl>, CMIG4.SA <dbl>,
#> #   CPFE3.SA <dbl>, CPLE6.SA <dbl>
#> 
#> $price.high
#> # A tibble: 248 x 16
#>    ref.date   ABEV3.SA BBAS3.SA BBDC3.SA BBDC4.SA BBSE3.SA BRAP4.SA BRFS3.SA
#>    <date>        <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#>  1 2014-01-02     17.5     24.5     22.0     20.1     24.5     25.1     49.2
#>  2 2014-01-03     17.4     24.2     21.7     19.9     24.0     24.2     49.0
#>  3 2014-01-06     17.1     24.0     21.4     19.9     23.7     24.0     48.0
#>  4 2014-01-07     17.4     24.1     21.6     19.9     23.3     24.0     48.2
#>  5 2014-01-08     17.6     23.6     21.2     19.7     23.3     23.6     47.4
#>  6 2014-01-09     17.5     23.0     20.9     19.4     23.2     23.2     45.8
#>  7 2014-01-10     17.2     23       20.8     19.3     23.1     22.6     45.4
#>  8 2014-01-13     17.3     23.1     20.9     19.3     23.2     22.6     45.0
#>  9 2014-01-14     17.3     23.5     20.5     19.1     23.5     22.4     45.5
#> 10 2014-01-15     17.2     23.5     20.7     19.3     24.1     22.5     45.2
#> # ... with 238 more rows, and 8 more variables: BRKM5.SA <dbl>, BRML3.SA <dbl>,
#> #   BVMF3.SA <dbl>, CCRO3.SA <dbl>, CIEL3.SA <dbl>, CMIG4.SA <dbl>,
#> #   CPFE3.SA <dbl>, CPLE6.SA <dbl>
#> 
#> $price.low
#> # A tibble: 248 x 16
#>    ref.date   ABEV3.SA BBAS3.SA BBDC3.SA BBDC4.SA BBSE3.SA BRAP4.SA BRFS3.SA
#>    <date>        <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#>  1 2014-01-02     17.2     24.0     21.6     19.8     23.7     24.0     48.0
#>  2 2014-01-03     16.9     23.8     21.0     19.6     23.3     23.6     46.9
#>  3 2014-01-06     16.5     23.6     21.1     19.5     22.9     23.5     47.0
#>  4 2014-01-07     16.9     23.4     21.0     19.4     22.6     23.1     47.1
#>  5 2014-01-08     17.2     23.0     20.8     19.2     22.8     23.2     45.8
#>  6 2014-01-09     17.0     22.1     20.3     18.8     22.4     21.9     44.5
#>  7 2014-01-10     16.9     22.1     20.2     18.8     22.5     22.1     44.0
#>  8 2014-01-13     17.0     22.4     20.2     18.7     22.9     22.1     44.1
#>  9 2014-01-14     17.0     22.5     20.2     18.8     23.0     22.0     44.2
#> 10 2014-01-15     16.9     23.0     20.4     19.0     23.5     22.2     44.4
#> # ... with 238 more rows, and 8 more variables: BRKM5.SA <dbl>, BRML3.SA <dbl>,
#> #   BVMF3.SA <dbl>, CCRO3.SA <dbl>, CIEL3.SA <dbl>, CMIG4.SA <dbl>,
#> #   CPFE3.SA <dbl>, CPLE6.SA <dbl>
#> 
#> $price.close
#> # A tibble: 248 x 16
#>    ref.date   ABEV3.SA BBAS3.SA BBDC3.SA BBDC4.SA BBSE3.SA BRAP4.SA BRFS3.SA
#>    <date>        <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#>  1 2014-01-02     17.4     24       21.7     19.8     24.0     24.1     48.5
#>  2 2014-01-03     17.0     24.2     21.2     19.7     23.4     24.1     47.2
#>  3 2014-01-06     16.8     23.9     21.4     19.9     23.1     23.8     47.2
#>  4 2014-01-07     17.3     23.5     21.0     19.4     23.0     23.2     47.5
#>  5 2014-01-08     17.5     23.0     20.9     19.4     23.2     23.3     45.9
#>  6 2014-01-09     17.1     22.2     20.3     18.9     22.7     22.2     44.6
#>  7 2014-01-10     17.2     23.0     20.6     19.0     23.0     22.2     44  
#>  8 2014-01-13     17.0     22.7     20.3     18.9     23.1     22.4     44.4
#>  9 2014-01-14     17.0     23.4     20.4     19.1     23.5     22.2     44.8
#> 10 2014-01-15     17.0     23.2     20.6     19.2     24.0     22.4     44.7
#> # ... with 238 more rows, and 8 more variables: BRKM5.SA <dbl>, BRML3.SA <dbl>,
#> #   BVMF3.SA <dbl>, CCRO3.SA <dbl>, CIEL3.SA <dbl>, CMIG4.SA <dbl>,
#> #   CPFE3.SA <dbl>, CPLE6.SA <dbl>
#> 
#> $volume
#> # A tibble: 248 x 16
#>    ref.date   ABEV3.SA BBAS3.SA BBDC3.SA BBDC4.SA BBSE3.SA BRAP4.SA BRFS3.SA
#>    <date>        <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#>  1 2014-01-02  8036139  4077900  1374911 10006894  4963400  1105600  2104600
#>  2 2014-01-03 24922793  5094600  5398069 19190794 13864500  1537500  6215300
#>  3 2014-01-06  9355961  3983200  1750418 12032869  5649900  1583300  1770600
#>  4 2014-01-07 18755025  4596200  1471816  7798256  5969300  1371000  2205800
#>  5 2014-01-08 11446953  6737300   940299 10682509  5974900   934500  6757400
#>  6 2014-01-09  8830612  8944700  1420007 11749439  7303900  2915100  4557400
#>  7 2014-01-10 13848974  8151600  1676279  9867356  4540500  1157200  2788100
#>  8 2014-01-13  5076667  4117500  1184604  8391689  4678100  1573700  1533900
#>  9 2014-01-14  7152058  4877200  1004367  9149197  4471300   983800  1414200
#> 10 2014-01-15  6173668  8992900   938255  7191611  5146300  1388500  1658400
#> # ... with 238 more rows, and 8 more variables: BRKM5.SA <dbl>, BRML3.SA <dbl>,
#> #   BVMF3.SA <dbl>, CCRO3.SA <dbl>, CIEL3.SA <dbl>, CMIG4.SA <dbl>,
#> #   CPFE3.SA <dbl>, CPLE6.SA <dbl>
#> 
#> $price.adjusted
#> # A tibble: 248 x 16
#>    ref.date   ABEV3.SA BBAS3.SA BBDC3.SA BBDC4.SA BBSE3.SA BRAP4.SA BRFS3.SA
#>    <date>        <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#>  1 2014-01-02     15.2     18.8     18.8     16.2     19.2     18.8     46.1
#>  2 2014-01-03     14.8     19.0     18.3     16.1     18.8     18.7     44.8
#>  3 2014-01-06     14.6     18.7     18.5     16.2     18.5     18.6     44.9
#>  4 2014-01-07     15.1     18.4     18.1     15.9     18.4     18.1     45.1
#>  5 2014-01-08     15.3     18.0     18.1     15.9     18.7     18.1     43.6
#>  6 2014-01-09     14.9     17.4     17.5     15.4     18.2     17.3     42.4
#>  7 2014-01-10     15.0     18.0     17.8     15.5     18.5     17.3     41.8
#>  8 2014-01-13     14.9     17.8     17.5     15.4     18.6     17.4     42.2
#>  9 2014-01-14     14.8     18.4     17.6     15.6     18.9     17.3     42.5
#> 10 2014-01-15     14.9     18.2     17.8     15.7     19.2     17.4     42.5
#> # ... with 238 more rows, and 8 more variables: BRKM5.SA <dbl>, BRML3.SA <dbl>,
#> #   BVMF3.SA <dbl>, CCRO3.SA <dbl>, CIEL3.SA <dbl>, CMIG4.SA <dbl>,
#> #   CPFE3.SA <dbl>, CPLE6.SA <dbl>
#> 
#> $ret.adjusted.prices
#> # A tibble: 248 x 16
#>    ref.date    ABEV3.SA BBAS3.SA  BBDC3.SA  BBDC4.SA  BBSE3.SA  BRAP4.SA
#>    <date>         <dbl>    <dbl>     <dbl>     <dbl>     <dbl>     <dbl>
#>  1 2014-01-02  NA        NA       NA        NA        NA        NA      
#>  2 2014-01-03  -0.0215    0.0104  -0.0224   -3.83e-3  -0.0220   -0.00332
#>  3 2014-01-06  -0.0130   -0.0150   0.00675   8.35e-3  -0.0142   -0.00877
#>  4 2014-01-07   0.0311   -0.0173  -0.0191   -2.21e-2  -0.00651  -0.0272 
#>  5 2014-01-08   0.00976  -0.0202  -0.00295  -7.10e-4   0.0126    0.00344
#>  6 2014-01-09  -0.0237   -0.0381  -0.0313   -2.91e-2  -0.0222   -0.0479 
#>  7 2014-01-10   0.00642   0.0355   0.0158    8.36e-3   0.0131    0      
#>  8 2014-01-13  -0.00935  -0.0110  -0.0151   -7.27e-3   0.00476   0.00763
#>  9 2014-01-14  -0.00294   0.0312   0.00774   9.08e-3   0.0146   -0.00628
#> 10 2014-01-15   0.00649  -0.0107   0.00668   5.77e-3   0.0202    0.00806
#> # ... with 238 more rows, and 9 more variables: BRFS3.SA <dbl>, BRKM5.SA <dbl>,
#> #   BRML3.SA <dbl>, BVMF3.SA <dbl>, CCRO3.SA <dbl>, CIEL3.SA <dbl>,
#> #   CMIG4.SA <dbl>, CPFE3.SA <dbl>, CPLE6.SA <dbl>
#> 
#> $ret.closing.prices
#> # A tibble: 248 x 16
#>    ref.date    ABEV3.SA BBAS3.SA  BBDC3.SA  BBDC4.SA  BBSE3.SA  BRAP4.SA
#>    <date>         <dbl>    <dbl>     <dbl>     <dbl>     <dbl>     <dbl>
#>  1 2014-01-02  NA        NA       NA        NA        NA        NA      
#>  2 2014-01-03  -2.15e-2   0.0104  -0.0233   -3.83e-3  -0.0220   -0.00332
#>  3 2014-01-06  -1.30e-2  -0.0150   0.00675   8.35e-3  -0.0142   -0.00877
#>  4 2014-01-07   3.11e-2  -0.0173  -0.0191   -2.21e-2  -0.00651  -0.0272 
#>  5 2014-01-08   9.76e-3  -0.0202  -0.00295  -7.10e-4   0.0126    0.00344
#>  6 2014-01-09  -2.37e-2  -0.0381  -0.0313   -2.91e-2  -0.0222   -0.0479 
#>  7 2014-01-10   6.42e-3   0.0355   0.0158    8.36e-3   0.0131    0      
#>  8 2014-01-13  -9.35e-3  -0.0110  -0.0151   -7.27e-3   0.00476   0.00763
#>  9 2014-01-14  -2.94e-3   0.0312   0.00774   9.08e-3   0.0146   -0.00628
#> 10 2014-01-15   5.88e-4  -0.0107   0.00668   5.77e-3   0.0202    0.00806
#> # ... with 238 more rows, and 9 more variables: BRFS3.SA <dbl>, BRKM5.SA <dbl>,
#> #   BRML3.SA <dbl>, BVMF3.SA <dbl>, CCRO3.SA <dbl>, CIEL3.SA <dbl>,
#> #   CMIG4.SA <dbl>, CPFE3.SA <dbl>, CPLE6.SA <dbl>
#>